(修正1)商品削除機能

+++ app/controllers/items_controller.rb
@@ -1,5 +1,6 @@
 class ItemsController < ApplicationController
-  before_action :authenticate_user!, only: [ :new, :create]
+  before_action :authenticate_user!, only: [ :new, :create, :destroy]
+  before_action :set_item, only: [:show, :destroy]
   
   def index 
     @items = Item.all.order("created_at DESC")
@@ -19,13 +20,15 @@ class ItemsController < ApplicationController
   end
 
   def show
-    @item = Item.find(params[:id])
   end
 
   def destroy
-    @item = Item.find(params[:id])
-    if @item.destroy
-      redirect_to root_path
+    if f @item.user_id == current_user.id
+      if @item.destroy
+        redirect_to root_path
+      end
+    else
+      render show
     end
   end
 
@@ -34,4 +37,9 @@ class ItemsController < ApplicationController
   def item_params
     params.require(:item).permit(:image, :name, :information, :category_id, :condition_id, :delivery_fee_id, :shipping_area_id, :shipping_day_id, :price).merge(user_id: current_user.id)
   end
+
+  def set_item
+    @item = Item.find(params[:id])
+  end
+
 end
